Rename registry to common
