+++ amazon/amazon.go
@@ -13,12 +13,12 @@ import (
 	"os"
 	"strconv"
 
-	"github.com/manics/binderhub-container-registry-helper/registry"
-
 	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/ecr"
 	"github.com/aws/aws-sdk-go-v2/service/ecr/types"
 	"github.com/aws/aws-sdk-go-v2/service/sts"
+
+	"github.com/manics/binderhub-container-registry-helper/common"
 )
 
 type IEcrClient interface {
@@ -53,13 +53,13 @@ func (c *ecrHandler) ListRepositories(w http.ResponseWriter, r *http.Request) {
 	repos, err := c.client.DescribeRepositories(context.TODO(), &input)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 	jsonBytes, err := json.Marshal(repos.Repositories)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 	w.WriteHeader(http.StatusOK)
@@ -93,7 +93,7 @@ func (c *ecrHandler) getRepoByName(name string) (*types.Repository, error) {
 func (c *ecrHandler) getRepositoryAsJson(r *http.Request) (bool, string, []byte, error) {
 	null := []byte("null")
 
-	name, err := registry.RepoGetName(r)
+	name, err := common.RepoGetName(r)
 	if err != nil {
 		return false, name, null, err
 	}
@@ -118,7 +118,7 @@ func (c *ecrHandler) GetRepository(w http.ResponseWriter, r *http.Request) {
 	found, name, jsonBytes, err := c.getRepositoryAsJson(r)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 
@@ -136,10 +136,10 @@ func (c *ecrHandler) GetRepository(w http.ResponseWriter, r *http.Request) {
 }
 
 func (c *ecrHandler) GetImage(w http.ResponseWriter, r *http.Request) {
-	repoName, tag, err := registry.ImageGetNameAndTag(r)
+	repoName, tag, err := common.ImageGetNameAndTag(r)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 	}
 	fullname := fmt.Sprintf("%s:%s", repoName, tag)
 
@@ -156,11 +156,11 @@ func (c *ecrHandler) GetImage(w http.ResponseWriter, r *http.Request) {
 		var awsErrRepo *types.RepositoryNotFoundException
 		if errors.As(err, &awsErrImage) || errors.As(err, &awsErrRepo) {
 			log.Printf("Repo '%s' not found\n", fullname)
-			registry.NotFound(w, r)
+			common.NotFound(w, r)
 			return
 		}
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 
@@ -169,7 +169,7 @@ func (c *ecrHandler) GetImage(w http.ResponseWriter, r *http.Request) {
 	jsonBytes, err := json.Marshal(image)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 	w.WriteHeader(http.StatusOK)
@@ -243,10 +243,10 @@ func (c *ecrHandler) setRepositoryPolicy(repoName string) error {
 }
 
 func (c *ecrHandler) CreateRepository(w http.ResponseWriter, r *http.Request) {
-	name, err := registry.RepoGetName(r)
+	name, err := common.RepoGetName(r)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 
@@ -268,19 +268,19 @@ func (c *ecrHandler) CreateRepository(w http.ResponseWriter, r *http.Request) {
 			found, _, jsonBytes, err := c.getRepositoryAsJson(r)
 			if err != nil {
 				log.Println("ERROR:", err)
-				registry.InternalServerError(w, r, err)
+				common.InternalServerError(w, r, err)
 				return
 			}
 			if !found {
 				log.Printf("RepositoryAlreadyExistsException but repository not found %s: %v", name, awsErr)
-				registry.InternalServerError(w, r, err)
+				common.InternalServerError(w, r, err)
 				return
 			}
 			log.Println("Repo already exists", name)
 			jsonResponse = jsonBytes
 		} else {
 			log.Println("ERROR:", err)
-			registry.InternalServerError(w, r, err)
+			common.InternalServerError(w, r, err)
 			return
 		}
 	}
@@ -288,7 +288,7 @@ func (c *ecrHandler) CreateRepository(w http.ResponseWriter, r *http.Request) {
 	err = c.setRepositoryPolicy(name)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 
@@ -296,7 +296,7 @@ func (c *ecrHandler) CreateRepository(w http.ResponseWriter, r *http.Request) {
 		jsonBytes, err := json.Marshal(createResponse.Repository)
 		if err != nil {
 			log.Println("ERROR:", err)
-			registry.InternalServerError(w, r, err)
+			common.InternalServerError(w, r, err)
 			return
 		}
 		jsonResponse = jsonBytes
@@ -331,10 +331,10 @@ func (c *ecrHandler) deleteRepositoryPolicy(repoName string) error {
 }
 
 func (c *ecrHandler) DeleteRepository(w http.ResponseWriter, r *http.Request) {
-	name, err := registry.RepoGetName(r)
+	name, err := common.RepoGetName(r)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 
@@ -343,7 +343,7 @@ func (c *ecrHandler) DeleteRepository(w http.ResponseWriter, r *http.Request) {
 	err = c.deleteRepositoryPolicy(name)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 
@@ -364,7 +364,7 @@ func (c *ecrHandler) DeleteRepository(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 
@@ -375,24 +375,24 @@ func (c *ecrHandler) GetToken(w http.ResponseWriter, r *http.Request) {
 	token, err := c.client.GetAuthorizationToken(context.TODO(), &ecr.GetAuthorizationTokenInput{})
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 	if len(token.AuthorizationData) != 1 {
 		msg := fmt.Errorf("expected 1 token, got %d", len(token.AuthorizationData))
 		log.Println("ERROR:", msg)
-		registry.InternalServerError(w, r, msg)
+		common.InternalServerError(w, r, msg)
 		return
 	}
 
-	resp := &registry.RegistryToken{
+	resp := &common.RegistryToken{
 		Token:   *token.AuthorizationData[0].AuthorizationToken,
 		Expires: *token.AuthorizationData[0].ExpiresAt,
 	}
 	jsonBytes, err := json.Marshal(resp)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 	w.WriteHeader(http.StatusOK)
@@ -417,7 +417,7 @@ func envvarIntGreaterThanZero(envvar string) (int, error) {
 	return i, nil
 }
 
-func Setup(args []string) (registry.IRegistryClient, error) {
+func Setup(args []string) (common.IRegistryClient, error) {
 	if len(args) != 0 {
 		return nil, errors.New("no arguments expected")
 	}
