+++ common/main_test.go
@@ -1,4 +1,4 @@
-package registry
+package common
 
 import (
 	"fmt"
