+++ go.mod
@@ -9,6 +9,7 @@ require (
 	github.com/aws/aws-sdk-go-v2/config v1.18.10
 	github.com/aws/aws-sdk-go-v2/service/ecr v1.18.1
 	github.com/aws/aws-sdk-go-v2/service/sts v1.18.2
+	github.com/oracle/oci-go-sdk v24.3.0+incompatible
 )
 
 require (
@@ -23,7 +24,6 @@ require (
 	github.com/aws/smithy-go v1.13.5 // indirect
 	github.com/gofrs/flock v0.8.1 // indirect
 	github.com/jmespath/go-jmespath v0.4.0 // indirect
-	github.com/oracle/oci-go-sdk v24.3.0+incompatible // indirect
 	github.com/sony/gobreaker v0.5.0 // indirect
 	golang.org/x/sys v0.4.0 // indirect
 )
@@ -32,4 +32,4 @@ replace github.com/manics/binderhub-container-registry-helper/oracle => ./oracle
 
 replace github.com/manics/binderhub-container-registry-helper/amazon => ./amazon
 
-replace github.com/manics/binderhub-container-registry-helper/registry => ../registry
+replace github.com/manics/binderhub-container-registry-helper/common => ../common
