+++ oracle/oracle.go
@@ -13,13 +13,13 @@ import (
 	"os"
 	"strings"
 
-	"github.com/manics/binderhub-container-registry-helper/registry"
-
 	"github.com/oracle/oci-go-sdk/identity"
 	"github.com/oracle/oci-go-sdk/v65/artifacts"
-	"github.com/oracle/oci-go-sdk/v65/common"
+	ocicommon "github.com/oracle/oci-go-sdk/v65/common"
 	"github.com/oracle/oci-go-sdk/v65/common/auth"
 	"github.com/oracle/oci-go-sdk/v65/objectstorage"
+
+	"github.com/manics/binderhub-container-registry-helper/common"
 )
 
 type IArtifactsClient interface {
@@ -45,13 +45,13 @@ func (c *artifactsHandler) ListRepositories(w http.ResponseWriter, r *http.Reque
 	})
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 	jsonBytes, err := json.Marshal(repos.Items)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 	w.WriteHeader(http.StatusOK)
@@ -75,7 +75,7 @@ func (c *artifactsHandler) dropNamespace(namespacedRepository string) (string, e
 }
 
 func (c *artifactsHandler) getByName(r *http.Request) (*artifacts.ContainerRepositorySummary, string, error) {
-	namespacedRepository, err := registry.RepoGetName(r)
+	namespacedRepository, err := common.RepoGetName(r)
 	if err != nil {
 		return nil, "", err
 	}
@@ -105,20 +105,20 @@ func (c *artifactsHandler) GetRepository(w http.ResponseWriter, r *http.Request)
 	repo, name, err := c.getByName(r)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 
 	log.Printf("Getting repo %s", name)
 
 	if repo == nil {
-		registry.NotFound(w, r)
+		common.NotFound(w, r)
 		return
 	} else {
 		jsonBytes, err := json.Marshal(repo)
 		if err != nil {
 			log.Println("ERROR:", err)
-			registry.InternalServerError(w, r, err)
+			common.InternalServerError(w, r, err)
 			return
 		}
 		w.WriteHeader(http.StatusOK)
@@ -130,15 +130,15 @@ func (c *artifactsHandler) GetRepository(w http.ResponseWriter, r *http.Request)
 }
 
 func (c *artifactsHandler) GetImage(w http.ResponseWriter, r *http.Request) {
-	namespacedRepository, tag, err := registry.ImageGetNameAndTag(r)
+	namespacedRepository, tag, err := common.ImageGetNameAndTag(r)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 	}
 	repoName, err := c.dropNamespace(namespacedRepository)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 	}
 
 	fullname := fmt.Sprintf("%s:%s", repoName, tag)
@@ -152,12 +152,12 @@ func (c *artifactsHandler) GetImage(w http.ResponseWriter, r *http.Request) {
 	})
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 	}
 
 	if len(images.Items) == 0 {
 		log.Printf("Image '%s' not found\n", fullname)
-		registry.NotFound(w, r)
+		common.NotFound(w, r)
 		return
 	}
 
@@ -166,7 +166,7 @@ func (c *artifactsHandler) GetImage(w http.ResponseWriter, r *http.Request) {
 	jsonBytes, err := json.Marshal(image)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 	w.WriteHeader(http.StatusOK)
@@ -177,16 +177,16 @@ func (c *artifactsHandler) GetImage(w http.ResponseWriter, r *http.Request) {
 }
 
 func (c *artifactsHandler) CreateRepository(w http.ResponseWriter, r *http.Request) {
-	namespacedRepository, err := registry.RepoGetName(r)
+	namespacedRepository, err := common.RepoGetName(r)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 	name, err := c.dropNamespace(namespacedRepository)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 
@@ -201,27 +201,27 @@ func (c *artifactsHandler) CreateRepository(w http.ResponseWriter, r *http.Reque
 
 	if err != nil {
 		log.Println("ERROR:", err)
-		serviceErr, ok := common.IsServiceError(err)
+		serviceErr, ok := ocicommon.IsServiceError(err)
 		if ok && serviceErr.GetCode() == "NAMESPACE_CONFLICT" {
 			log.Printf("Repository already exists: %v\n", err)
 
 			repo, name, err := c.getByName(r)
 			if err != nil {
 				log.Println("ERROR:", err)
-				registry.InternalServerError(w, r, err)
+				common.InternalServerError(w, r, err)
 				return
 			}
 
 			if repo == nil {
 				log.Printf("NAMESPACE_CONFLICT but repository not found %s: %v", name, err)
-				registry.InternalServerError(w, r, err)
+				common.InternalServerError(w, r, err)
 				return
 			}
 
 			jsonBytes, err := json.Marshal(repo)
 			if err != nil {
 				log.Println("ERROR:", err)
-				registry.InternalServerError(w, r, err)
+				common.InternalServerError(w, r, err)
 				return
 			}
 
@@ -232,7 +232,7 @@ func (c *artifactsHandler) CreateRepository(w http.ResponseWriter, r *http.Reque
 			}
 			return
 		} else {
-			registry.InternalServerError(w, r, err)
+			common.InternalServerError(w, r, err)
 			return
 		}
 	}
@@ -240,7 +240,7 @@ func (c *artifactsHandler) CreateRepository(w http.ResponseWriter, r *http.Reque
 	jsonBytes, err := json.Marshal(createResponse.ContainerRepository)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 	w.WriteHeader(http.StatusOK)
@@ -254,7 +254,7 @@ func (c *artifactsHandler) DeleteRepository(w http.ResponseWriter, r *http.Reque
 	repo, name, err := c.getByName(r)
 	if err != nil {
 		log.Println("ERROR:", err)
-		registry.InternalServerError(w, r, err)
+		common.InternalServerError(w, r, err)
 		return
 	}
 
@@ -266,7 +266,7 @@ func (c *artifactsHandler) DeleteRepository(w http.ResponseWriter, r *http.Reque
 		})
 		if err != nil {
 			log.Println("ERROR:", err)
-			registry.InternalServerError(w, r, err)
+			common.InternalServerError(w, r, err)
 			return
 		}
 		w.WriteHeader(http.StatusOK)
@@ -276,11 +276,11 @@ func (c *artifactsHandler) DeleteRepository(w http.ResponseWriter, r *http.Reque
 
 func (c *artifactsHandler) GetToken(w http.ResponseWriter, r *http.Request) {
 	log.Println("GetToken not implemented")
-	registry.NotFound(w, r)
+	common.NotFound(w, r)
 }
 
-func Setup(args []string) (registry.IRegistryClient, error) {
-	var cfg common.ConfigurationProvider
+func Setup(args []string) (common.IRegistryClient, error) {
+	var cfg ocicommon.ConfigurationProvider
 	var err error
 
 	switch len(args) {
@@ -295,7 +295,7 @@ func Setup(args []string) (registry.IRegistryClient, error) {
 	case 1:
 		// User principals, using configuration file
 		cfg_file := args[0]
-		cfg, err = common.ConfigurationProviderFromFile(cfg_file, "")
+		cfg, err = ocicommon.ConfigurationProviderFromFile(cfg_file, "")
 		if err != nil {
 			log.Printf("failed to load configuration, %v", err)
 			return nil, err
